IGNITE-7816: DataFrame examples moved to `org.apache.ignite.scalar.examples.spark` package. Minor examples improvements.
