+++ examples/src/main/scala/org/apache/ignite/scalar/examples/spark/package.scala
@@ -14,7 +14,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.ignite.examples
+
+package org.apache.ignite.scalar.examples
 
 /**
   */
