+++ examples/src/test/spark/org/apache/ignite/spark/examples/IgniteDataFrameSelfTest.java
@@ -17,14 +17,15 @@
 
 package org.apache.ignite.spark.examples;
 
-import org.apache.ignite.examples.spark.IgniteCatalogExample;
-import org.apache.ignite.examples.spark.IgniteDataFrameExample;
-import org.apache.ignite.examples.spark.IgniteDataFrameWriteExample;
+import org.apache.ignite.scalar.examples.spark.IgniteCatalogExample;
+import org.apache.ignite.scalar.examples.spark.IgniteDataFrameExample;
+import org.apache.ignite.scalar.examples.spark.IgniteDataFrameWriteExample;
+import org.apache.ignite.testframework.junits.common.GridAbstractExamplesTest;
 import org.junit.Test;
 
 /**
  */
-public class IgniteDataFrameSelfTest {
+public class IgniteDataFrameSelfTest extends GridAbstractExamplesTest {
     static final String[] EMPTY_ARGS = new String[0];
 
     /**
